clippy & others
