+++ .github/workflows/ci.yml
@@ -149,7 +149,6 @@ jobs:
 
       - name: Run other tests
         run: |
-          cargo test --test request_reply_test --test memory_test --features full -- --ignored --nocapture --test-threads=1
           cargo test --test integration_test --features full -- --ignored --nocapture --test-threads=1 --skip test_all_chaos --skip test_all_performance_pipeline
 
       - name: Run armature integration tests
