+++ tests/integration/mod.rs
@@ -20,4 +20,5 @@ pub mod zeromq;
 pub mod file;
 pub mod logic_test;
 pub mod memory;
+// request_reply_test.rs merged into route.rs
 pub mod route;
