+++ tests/integration_test.rs
@@ -5,8 +5,6 @@
 
 #[path = "integration/mod.rs"]
 mod integration;
-mod memory_test;
-mod request_reply_test;
 
 #[allow(dead_code)]
 pub fn should_run(test_name: &str) -> bool {
@@ -268,4 +266,21 @@ async fn test_all_performance_pipeline() {
             integration::zeromq::test_zeromq_performance_pipeline().await;
         }
     }
+}
+
+#[tokio::test(flavor = "multi_thread")]
+#[ignore = "requires docker compose"]
+async fn test_all_request_reply() {
+    println!("--- Running All Request-Reply Tests ---");
+    #[cfg(feature = "kafka")]
+    integration::route::test_kafka_request_reply().await;
+    #[cfg(feature = "nats")]
+    integration::route::test_nats_request_reply().await;
+    #[cfg(feature = "mongodb")]
+    integration::route::test_mongodb_request_reply_pattern().await;
+    #[cfg(feature = "amqp")]
+    integration::route::test_amqp_request_reply().await;
+    #[cfg(feature = "mqtt")]
+    integration::route::test_mqtt_request_reply().await;
+    integration::route::test_memory_request_reply().await;
 }
